+++ lib/export.sh
@@ -6,14 +6,14 @@ export_aliases() {
   ensure_store
   while IFS= read -r name; do
     read_alias "$name"
-    local meta="# alx:name=${name}"
-    if [[ -n $_ALX_DESC ]]; then
-      meta+=" desc=\"$(escape_meta_desc "$_ALX_DESC")\""
+    local comment=""
+    if [[ -n $_ALX_DESC && -n $_ALX_TAGS ]]; then
+      comment=" # $_ALX_DESC [${_ALX_TAGS//,/, }]"
+    elif [[ -n $_ALX_DESC ]]; then
+      comment=" # $_ALX_DESC"
+    elif [[ -n $_ALX_TAGS ]]; then
+      comment=" # [${_ALX_TAGS//,/, }]"
     fi
-    if [[ -n $_ALX_TAGS ]]; then
-      meta+=" tags=${_ALX_TAGS}"
-    fi
-    printf '%s\n' "$meta"
-    printf "alias -- %s='%s'\n" "$name" "$(shell_escape_single "$_ALX_CMD")"
+    printf "alias %s='%s'%s\n" "$name" "$(shell_escape_single "$_ALX_CMD")" "$comment"
   done < <(list_alias_names)
 }